requirements.txt
tensorflow>=2.15.0
numpy>=1.24.3
matplotlib>=3.7.1
seaborn>=0.12.2
opencv-python>=4.8.0
scikit-learn>=1.3.0
pandas>=2.0.3
streamlit>=1.32.0
Pillow>=10.0.0
plotly>=5.18.0
